requirements.txt
pandas
python-dotenv
tqdm
openai>=1.30.0
transformers>=4.41.0
torch --extra-index-url https://download.pytorch.org/whl/cpu
google-cloud-texttospeech
pydub
soundfile
numpy
